Update a few lines on MyBot.py
Replace the MyBot.py from your starter bot with this file and it works slightly better. I added an array list at the top that the ships use to designate which planets they're all going to. If the planet is already in the list, the ship decides to go to another planet. I also doubled the speed of the ships.